Extracted a more generic execute command
