+++ lib/runTarget.js
@@ -1,50 +1,29 @@
-const {
-  pipeline,
-  forEach,
-  flatMap,
-  parallel,
-} = require("@transformation/core");
-const { execFile } = require("child_process");
+const { filter, pipeline } = require("@transformation/core");
 const chalk = require("chalk");
-const util = require("util");
-const execFileAsync = util.promisify(execFile);
-const spinner = require("./spinner");
+const execute = require("./execute");
 
 const emphasized = text => chalk.cyan.bold(text);
 
-const runTarget = ({ npmClient, args, concurrency }) =>
-  pipeline(
-    parallel(
-      flatMap(async ({ name, packageDefinition, dir }) => {
-        const target = args[0];
-        const heading = `${name}: ${npmClient} ${args.join(" ")}`;
+const runTarget = ({ npmClient, args, concurrency }) => {
+  const target = args[0];
 
-        if (packageDefinition.scripts && packageDefinition.scripts[target]) {
-          try {
-            const { stdout, stderr } = await execFileAsync(npmClient, args, {
-              cwd: dir,
-              env: { FORCE_COLOR: "true", ...process.env },
-            });
+  return pipeline(
+    filter(({ name, packageDefinition }) => {
+      const hasTarget =
+        packageDefinition.scripts && packageDefinition.scripts[target];
 
-            return [emphasized(heading), stdout, stderr].join("\n");
-          } catch (e) {
-            process.stderr.write(
-              emphasized(`${heading} (${chalk.red("failed")})\n`)
-            );
-            process.stderr.write(e.message);
-            process.stderr.write("\n");
-            throw e;
-          }
-        } else {
-          return [emphasized(`${heading} (skip)\n`)];
-        }
-      }),
-      concurrency
-    ),
-    spinner("Waiting for input"),
-    forEach(message => {
-      process.stdout.write(message);
-    })
+      if (!hasTarget) {
+        process.stderr.write(
+          emphasized(
+            emphasized(`${name}: ${npmClient} ${args.join(" ")} (skip)\n`)
+          )
+        );
+      }
+
+      return hasTarget;
+    }),
+    execute({ command: "yarn", args, concurrency })
   );
+};
 
 module.exports = runTarget;
